+++ client/src/lib/constants.ts
@@ -54,3 +54,42 @@ export const DEFAULT_INSPECTOR_CONFIG: InspectorConfig = {
     value: "",
   },
 } as const;
+
+interface ClaudeModel {
+  id: string;
+  name: string;
+  description: string;
+}
+
+export const CLAUDE_MODELS: ClaudeModel[] = [
+  {
+    id: "claude-opus-4-0",
+    name: "Claude Opus 4",
+    description: "Latest and most powerful model for complex reasoning",
+  },
+  {
+    id: "claude-sonnet-4-0",
+    name: "Claude Sonnet 4",
+    description: "Next generation balanced model with enhanced capabilities",
+  },
+  {
+    id: "claude-3-7-sonnet-latest",
+    name: "Claude Sonnet 3.7",
+    description: "Most intelligent model with extended thinking",
+  },
+  {
+    id: "claude-3-5-sonnet-latest",
+    name: "Claude Sonnet 3.5",
+    description: "High level of intelligence and capability",
+  },
+  {
+    id: "claude-3-5-haiku-latest",
+    name: "Claude Haiku 3.5",
+    description: "Fastest model - intelligence at blazing speeds",
+  },
+  {
+    id: "claude-3-opus-latest",
+    name: "Claude Opus 3",
+    description: "Top-level intelligence, fluency, and understanding",
+  },
+];
